django/prac/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
ChaCha20==1.1.1
charset-normalizer==3.2.0
cryptography==41.0.4
Django==4.2.6
idna==3.4
numpy==1.26.0
pandas==2.1.1
paytm-pg==1.1.1
paytmchecksum==1.7.0
pycparser==2.21
pycryptodome==3.19.0
python-dateutil==2.8.2
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
asgiref==3.7.2
beautifulsoup4==4.12.2
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
soupsieve==2.5

